+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Update the oldest supported Emacs version to 27.1.
 * Add "MOVE AROUND THINGs" section to `meow-tutor.el`.
+* Update `meow-goto-line` to expand `meow-line`.
 
 ### Bugs fixed